+++ pkg/webhook/secrets.go
@@ -23,7 +23,8 @@ import (
 )
 
 // Vault Sidecar Injector: Secrets Mode
-func (vaultInjector *VaultInjector) secretsMode(labels, annotations map[string]string) (string, error) {
+func (vaultInjector *VaultInjector) secretsMode(labels, annotations map[string]string) (modeConfig, error) {
+	secretsType := annotations[vaultInjector.VaultInjectorAnnotationsFQ[vaultInjectorAnnotationSecretsTypeKey]]
 	secretsPath := strings.Split(annotations[vaultInjector.VaultInjectorAnnotationsFQ[vaultInjectorAnnotationSecretsPathKey]], ",")
 	secretsTemplate := strings.Split(annotations[vaultInjector.VaultInjectorAnnotationsFQ[vaultInjectorAnnotationSecretsTemplateKey]], "---")
 	templateDest := strings.Split(annotations[vaultInjector.VaultInjectorAnnotationsFQ[vaultInjectorAnnotationTemplateDestKey]], ",")
@@ -33,13 +34,32 @@ func (vaultInjector *VaultInjector) secretsMode(labels, annotations map[string]s
 	secretsTemplateNum := len(secretsTemplate)
 	templateDestNum := len(templateDest)
 
+	if secretsType == "" {
+		secretsType = vaultInjectorSecretsTypeDynamic // Dynamic secrets by default
+	} else {
+		secretsTypeSupported := false
+		for _, supportedSecretType := range vaultInjectorSecretsTypes {
+			if secretsType == supportedSecretType {
+				secretsTypeSupported = true
+				break
+			}
+		}
+
+		if !secretsTypeSupported {
+			err := fmt.Errorf("Submitted pod makes use of unsupported secrets type %s", secretsType)
+			klog.Error(err.Error())
+			return nil, err
+		}
+	}
+
 	if secretsPathNum == 1 && secretsPath[0] == "" { // Build default secrets path: "secret/<application label>/<service label>"
 		applicationLabel := labels[vaultInjector.ApplicationLabelKey]
 		applicationServiceLabel := labels[vaultInjector.ApplicationServiceLabelKey]
 
 		if applicationLabel == "" || applicationServiceLabel == "" {
-			klog.Errorf("Submitted pod must contain labels %s and %s", vaultInjector.ApplicationLabelKey, vaultInjector.ApplicationServiceLabelKey)
-			return "", fmt.Errorf("Submitted pod must contain labels %s and %s", vaultInjector.ApplicationLabelKey, vaultInjector.ApplicationServiceLabelKey)
+			err := fmt.Errorf("Submitted pod must contain labels %s and %s", vaultInjector.ApplicationLabelKey, vaultInjector.ApplicationServiceLabelKey)
+			klog.Error(err.Error())
+			return nil, err
 		}
 
 		secretsPath[0] = vaultDefaultSecretsEnginePath + "/" + applicationLabel + "/" + applicationServiceLabel
@@ -52,8 +72,9 @@ func (vaultInjector *VaultInjector) secretsMode(labels, annotations map[string]s
 	if secretsTemplateNum == 1 && secretsTemplate[0] == "" {
 		// We must have same numbers of secrets path & secrets destinations
 		if templateDestNum != secretsPathNum {
-			klog.Error("Submitted pod must contain same numbers of secrets path and secrets destinations")
-			return "", errors.New("Submitted pod must contain same numbers of secrets path and secrets destinations")
+			err := errors.New("Submitted pod must contain same numbers of secrets path and secrets destinations")
+			klog.Error(err.Error())
+			return nil, err
 		}
 
 		// If no custom template(s), use default template
@@ -64,8 +85,9 @@ func (vaultInjector *VaultInjector) secretsMode(labels, annotations map[string]s
 	} else {
 		// We must have same numbers of custom templates & secrets destinations ...
 		if templateDestNum != secretsTemplateNum {
-			klog.Error("Submitted pod must contain same numbers of templates and secrets destinations")
-			return "", errors.New("Submitted pod must contain same numbers of templates and secrets destinations")
+			err := errors.New("Submitted pod must contain same numbers of templates and secrets destinations")
+			klog.Error(err.Error())
+			return nil, err
 		}
 
 		// ... and we ignore content of 'secrets-path' annotation ('cause we provide full template), but we need to init an empty array
@@ -91,5 +113,21 @@ func (vaultInjector *VaultInjector) secretsMode(labels, annotations map[string]s
 		templates.WriteString("\n")
 	}
 
-	return templates.String(), nil
+	return &secretsModeConfig{secretsType, templates.String()}, nil
+}
+
+func (secModeCfg *secretsModeConfig) getTemplate() string {
+	return secModeCfg.template
+}
+
+func getSecretsType(config modeConfig) (string, error) {
+	secModeCfg, ok := config.(*secretsModeConfig) // here we use type assertion (https://golang.org/ref/spec#Type_assertions)
+
+	if ok {
+		return secModeCfg.secretsType, nil
+	}
+
+	err := errors.New("Provided type cannot be casted to 'secretsModeConfig'")
+	klog.Error(err.Error())
+	return "", err
 }